Call onMoveCallBack at the beginning
A common way to know if a drag is occurring is to compate the
mouse status with the previous mouse status.

This can be used to register where a move started from:

```js
if(previous!=='down' && current === 'down'){
  registerMousePosition()
}
```

This is useful to draw straight lines in a drawing app, for example.

Without `onMoveCallback` running at least once before, there's no way
of knowing when the movement has started.